refactor: split SpotifyContext into reducer and context
Extract reducer, action types, state interface, and initial state
into spotifyReducer.ts. SpotifyContext.tsx now only contains the
provider, hooks, and API callback wrappers.
